Add new tutorial for deploying React app to AWS
A comprehensive guide for deploying a React application to AWS Cloudfront using Hereya has been added. This includes the process of creating a new React app with Vite, initializing Hereya within the project, and deploying the app. Detailed steps, necessary prerequisites and commands are given. Steps for clean up after completing the tutorial are also provided.